+++ src/test/kotlin/com/terraformation/backend/seedbank/db/PhotoRepositoryTest.kt
@@ -89,6 +89,7 @@ class PhotoRepositoryTest : DatabaseTest(), RunsAsUser {
             mockk(),
             clock,
             mockk(),
+            mockk(),
         )
 
     tempDir = Files.createTempDirectory(javaClass.simpleName)
